installed retry hack to compensate for infrequent race conditions in RaceCapture/Pro USB-Comm. remove this retry mechanism when resolved in firmware

+++ src/comm.cpp
@@ -171,7 +171,7 @@ void RaceAnalyzerComm::FlushReceiveBuffer(CComm* comPort){
 	comPort->drainInput();
 }
 
-wxString RaceAnalyzerComm::SendCommand(CComm *comPort, const wxString &buffer, int timeout){
+wxString RaceAnalyzerComm::SendCommand(CComm *comPort, const wxString &buffer, size_t timeout){
 
 	wxString response;
